requirements.txt
tensorflow>=2.15,<3.0
numpy>=1.24
pandas>=2.0
matplotlib>=3.8
scikit-learn>=1.3
umap-learn>=0.5
pillow>=10.0
jupyter>=1.0
pytest>=8.0
black>=24.0
flake8>=7.0
pylint>=3.0
